+++ NEWS.md
@@ -1,6 +1,9 @@
 # agridat 1.27 ()
 
+## New datasets
+
 * clewer.wheat
+* desousa.potato
 * petersen.wheat
 
 # agridat 1.26 (2025-11-14)
@@ -9,7 +12,6 @@
 
 * kirk.potato.uniformity
 * wehner.cucumber.uniformity
-
 
 # agridat 1.25 (2025-10-06)